Tests/LibWeb/Text/expected/Streams/ReadableStreamBYOBReader-read-non-transferable-buffer.txt

read() rejected with TypeError: true
Read buffer byteLength: 65536
enqueue() threw TypeError: true
Enqueue buffer byteLength: 65536
respondWithNewView() rejected with TypeError: true
RespondWithNewView buffer byteLength: 65536
